Teach cygwin-console-helper to output kernel32!CtrlRoutine's  address
This will help us with emulating SIGINT properly (by not sending signals
to *all* processes attached to the same Console).

We will extend this functionality to handle the address of
ExitProcess(), too. Hence we ask which function's address to print.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

squash! Teach cygwin-console-helper to output the address of kernel32!CtrlRoutine