pipe: fix SSH hang (again)
The recent changes in Cygwin's pipe logic are a gift that keeps on
giving.

The recurring pattern is that bugs are fixed, but with many of these bug
fixes, new bugs are produced, and we are stuck in this seemingly endless
tunnel of fixing bugs caused by bug fixes.

In cbfaeba4f7 (Cygwin: pipe: Fix incorrect write length in raw_write(),
2024-11-06), for example, a segmentation fault was fixed. Which is good!
However, a bug was introduced, where e.g. Git for Windows' `git clone`
via SSH frequently hangs indefinitely for large-ish repositories, see
https://github.com/git-for-windows/git/issues/5688.

That commit removed logic whereby in non-blocking mode, not only the
chunk size, but also the overall count of bytes to write (`len`) was
clamped to whatever is indicated as the `WriteQuotaAvailable`. Now only
`chunk` is clamped to that, but `len` is left at its original number.
However, the following `while` loop expects `len - chunk` (which is
assigned to `len1`) not to be positive in non-blocking mode.

Let's reinstate that clamping logic and implicitly fix this SSH hang.

Fixes: cbfaeba4f7 (Cygwin: pipe: Fix incorrect write length in raw_write(), 2024-11-06)
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>